+++ commands/src/main/java/org/wildfly/extras/creaper/commands/datasources/AddPostgresPlusDataSource.java
@@ -30,7 +30,7 @@ public final class AddPostgresPlusDataSource extends AddDataSource {
         if (exceptionSorterClass == null) exceptionSorterClass = POSTGRES_PLUS_EXCEPTION_SORTER;
     }
 
-    public static final class Builder extends AddDataSource.Builder {
+    public static final class Builder extends AddDataSource.Builder<Builder> {
         public Builder(String name) {
             super(name);
         }