+++ commands/src/main/java/org/wildfly/extras/creaper/commands/datasources/AddSybaseDataSource.java
@@ -28,7 +28,7 @@ public final class AddSybaseDataSource extends AddDataSource {
         if (exceptionSorterClass == null) exceptionSorterClass = SYBASE_EXCEPTION_SORTER;
     }
 
-    public static final class Builder extends AddDataSource.Builder {
+    public static final class Builder extends AddDataSource.Builder<Builder> {
         public Builder(String name) {
             super(name);
         }