+++ commands/src/main/java/org/wildfly/extras/creaper/commands/datasources/AddSybaseXADataSource.java
@@ -45,7 +45,7 @@ public final class AddSybaseXADataSource extends AddXADataSource {
         if (!xaDatasourceProperties.containsKey("PortNumber")) xaDatasourceProperties.put("PortNumber", "4100");
     }
 
-    public static final class Builder extends AddXADataSource.Builder {
+    public static final class Builder extends AddXADataSource.Builder<Builder> {
         public Builder(String name) {
             super(name);
         }